+++ CHANGELOG.md
@@ -6,6 +6,7 @@ UNRELEASED
 * Fix page permissions
 * Show recurrence in event list
 * Only show upcoming events per default
+* Allow configuration via /etc/integreat-cms.ini
 
 
 2021.11.0-beta